Move code around
# Conflicts:
#	src/Microsoft.DotNet.XHarness.CLI/CommandArguments/iOS/iOSPackageCommandArguments.cs
#	src/Microsoft.DotNet.XHarness.CLI/Commands/XHarnessCommand.cs
#	src/Microsoft.DotNet.XHarness.CLI/Commands/iOS/iOSPackageCommand.cs
#	src/Microsoft.DotNet.XHarness.CLI/Microsoft.DotNet.XHarness.CLI.csproj
