+++ src/Microsoft.DotNet.XHarness.CLI/Commands/Android/AndroidCommandSet.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using Microsoft.DotNet.XHarness.CLI.Commands.Android;
 using Mono.Options;
 
 namespace Microsoft.DotNet.XHarness.CLI.Android
