+++ src/Microsoft.DotNet.XHarness.CLI/Commands/XHarnessCommand.cs
@@ -1,4 +1,4 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
+// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
@@ -7,10 +7,11 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Microsoft.DotNet.XHarness.CLI.CommandArguments;
 using Microsoft.Extensions.Logging;
 using Mono.Options;
 
-namespace Microsoft.DotNet.XHarness.CLI.Common
+namespace Microsoft.DotNet.XHarness.CLI.Commands
 {
     internal abstract class XHarnessCommand : Command
     {
@@ -70,7 +71,7 @@ namespace Microsoft.DotNet.XHarness.CLI.Common
                     return 1;
                 }
 
-                return (int) InvokeInternal().GetAwaiter().GetResult();
+                return (int)InvokeInternal().GetAwaiter().GetResult();
             }
             finally
             {
